pulumi/requirements.txt
pulumi>=3.0.0,<4.0.0
pulumi-aws>=6.0.2,<7.0.0
PyYAML>=6.0.1

# dev requirements
# TODO: separate dev requirements
ruff>=0.3.5
